requirements.txt
pandas>=2.0.0
pyJWT>=2.8.0
chromadb>=0.5.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.9.0
openai>=1.30.0
sentence-transformers>=3.0.0
supabase>=2.4.0
bcrypt>=4.1.0
pydantic[email]>=2.7.0
